+++ haproxy-operator/src/haproxy.py
@@ -130,6 +130,7 @@ class HAProxyService:
         """
         template_context = {
             "config_global_max_connection": charm_state.global_max_connection,
+            "disable_ddos_protection": charm_state.disable_ddos_protection,
             "ingress_requirers_information": ingress_requirers_information,
             "config_external_hostname": external_hostname,
             "haproxy_crt_dir": HAPROXY_CERTS_DIR,
@@ -159,6 +160,7 @@ class HAProxyService:
         """
         template_context = {
             "config_global_max_connection": charm_state.global_max_connection,
+            "disable_ddos_protection": charm_state.disable_ddos_protection,
             "backends": haproxy_route_requirers_information.backends,
             "tcp_endpoints": haproxy_route_requirers_information.tcp_endpoints,
             "stick_table_entries": haproxy_route_requirers_information.stick_table_entries,
@@ -194,6 +196,7 @@ class HAProxyService:
             HAPROXY_DEFAULT_CONFIG_TEMPLATE,
             {
                 "config_global_max_connection": charm_state.global_max_connection,
+                "disable_ddos_protection": charm_state.disable_ddos_protection,
             },
         )
         self._validate_haproxy_config()