+++ docs/changelog.md
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 Each revision is versioned by the date of the revision.
 
+## 2025-12-17
+
+- Added the DDoS defaults in the HAProxy config files and a config option to disable the defaults.
+
 ## 2025-12-11
 
 - Added the skeleton code for the HAProxy DDoS protection configurator charm.